EZP-31651: Decoupled Search Bundle from AdminUI (#1404)
* EZP-31651: Moved user_content_type_identifier to adminUI

* EZP-31651: Move and mark classes as deprecated

* EZP-31651: Moved deprecacation time to version 3.2

* EZP-31651: CS and DI fix
