+++ src/bundle/Resources/config/services/query_types.yaml
@@ -11,5 +11,6 @@ services:
     EzSystems\EzPlatformAdminUi\QueryType\MediaSubtreeQueryType: ~
 
     EzSystems\EzPlatformAdminUi\QueryType\SearchQueryType:
+        deprecated: 'Service "%service_id%" is deprecated since 3.1 and will be removed in 3.2. Use "Ibexa\Platform\Search\QueryType\SearchQueryType" instead'
         arguments:
             $searchService: '@ezpublish.api.service.search'
