+++ src/lib/EventListener/SearchListViewFilterParametersListener.php
@@ -6,23 +6,42 @@
  */
 namespace EzSystems\EzPlatformAdminUi\EventListener;
 
+use eZ\Publish\Core\MVC\ConfigResolverInterface;
 use eZ\Publish\Core\MVC\Symfony\View\Event\FilterViewParametersEvent;
 use eZ\Publish\Core\MVC\Symfony\View\ViewEvents;
 use EzSystems\EzPlatformAdminUi\Form\Data\Content\Draft\ContentEditData;
 use EzSystems\EzPlatformAdminUi\Form\Type\Content\Draft\ContentEditType;
+use EzSystems\EzPlatformAdminUi\Specification\SiteAccess\IsAdmin;
 use Ibexa\Platform\Search\View\SearchListView;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Form\FormFactoryInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestStack;
 
-final class EditFormSearchListViewFilterParametersListener implements EventSubscriberInterface
+final class SearchListViewFilterParametersListener implements EventSubscriberInterface
 {
     /** @var \Symfony\Component\Form\FormFactoryInterface */
     private $formFactory;
 
+    /** @var \eZ\Publish\Core\MVC\ConfigResolverInterface */
+    private $configResolver;
+
+    /** @var \Symfony\Component\HttpFoundation\RequestStack */
+    private $requestStack;
+
+    /** @var string[][] */
+    private $siteAccessGroups;
+
     public function __construct(
-        FormFactoryInterface $formFactory
+        FormFactoryInterface $formFactory,
+        ConfigResolverInterface $configResolver,
+        RequestStack $requestStack,
+        array $siteAccessGroups
     ) {
         $this->formFactory = $formFactory;
+        $this->configResolver = $configResolver;
+        $this->requestStack = $requestStack;
+        $this->siteAccessGroups = $siteAccessGroups;
     }
 
     public static function getSubscribedEvents(): array
@@ -40,6 +59,10 @@ final class EditFormSearchListViewFilterParametersListener implements EventSubsc
             return;
         }
 
+        if (!$this->isAdminSiteAccess($this->requestStack->getCurrentRequest())) {
+            return;
+        }
+
         $editForm = $this->formFactory->create(
             ContentEditType::class,
             new ContentEditData(),
@@ -47,6 +70,12 @@ final class EditFormSearchListViewFilterParametersListener implements EventSubsc
 
         $event->getParameterBag()->add([
             'form_edit' => $editForm->createView(),
+            'user_content_type_identifier' => $this->configResolver->getParameter('user_content_type_identifier'),
         ]);
+    }
+
+    private function isAdminSiteAccess(Request $request): bool
+    {
+        return (new IsAdmin($this->siteAccessGroups))->isSatisfiedBy($request->attributes->get('siteaccess'));
     }
 }
